chore: separated documenation read from write
