+++ libflux/flux/src/lib.rs
@@ -4,7 +4,10 @@ pub mod wasm;
 
 extern crate fluxcore;
 extern crate serde_aux;
+
+#[macro_use]
 extern crate serde_derive;
+pub mod docs;
 
 use fluxcore::parser::Parser;
 use fluxcore::semantic::check;
